requirements.txt
pyfiglet==0.8.post1
validictory==1.1.2
google_auth_oauthlib==0.4.1
tabulate==0.8.7
google_api_python_client==1.9.3
mysql-connector-python==8.0.20
PyInquirer==1.0.3
python_dateutil==2.8.1
